requirements-cloudrun-gpu.txt
# Palm Counter — Cloud Run GPU production requirements
# Used by Dockerfile.gpu (commercial tier, NVIDIA L4 / CUDA 12.2)
#
# Key difference from requirements.txt:
#   onnxruntime        → onnxruntime-gpu
#   (app/core/inference.py auto-selects CUDAExecutionProvider at import time)
#
# Do NOT use this file for local CPU-only development.
# Use requirements.txt (CPU) or requirements-cuda.txt (local GPU dev).

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Image & raster
Pillow>=10.0.0
tifffile>=2024.1.1
rasterio>=1.3.0
numpy>=1.26.0

# YOLO / ONNX inference — GPU-accelerated
# CUDA 12.2 on Cloud Run L4 requires onnxruntime-gpu >= 1.18.0
opencv-python-headless>=4.9.0
onnxruntime-gpu>=1.18.0

# Geo
geopandas>=0.14.0
shapely>=2.0.0
rtree>=1.0.0
pyproj>=3.6.0
topojson>=1.9        # topology-preserving simplification for land cover polygons

# Utilities
pyyaml>=6.0
tqdm>=4.66.0

# Auth & cloud config
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
pydantic-settings>=2.0.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
